requirement.txt
lxml==4.3.2
pyyaml==5.1.2
PyQt5==5.13.0